pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Free check sites for SEO suitability"
authors = ["Andrey Bogatyrev <t2way@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
classifiers = ["Programming Language :: Python :: 3",
               "Environment :: Web"]
repository = "https://github.com/Onoiro/python-project-83"


[tool.poetry.scripts]
page-analyzer = "page_analyzer.app:app"


[tool.poetry.dependencies]
python = "^3.8.1"
flask = "^2.3.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
